Decouple Python SDK from its repository
Originally, we included the Python SDK in the requirements. However,
this tightly couples C# SDK with Python SDK so we cannot update either
of the repositories to different versions of aas-core-meta.

We copy now the Python code directly from aas-core-codegen, and decouple
thus this repository from aas-core3.0-python. This allows us now to
atomically update *this repository* to the latest aas-core-meta without
touching aas-core3.0-python.