Refactor unwrapper from enhancer
Currently, you can only specify the unwrapping if you also specify the
wrapping. This is unfortunate, as wrapping might occur in a different
part of the code base from where you define the unwrapping.

In particular, the enhancer needs an enhancement factory -- which might
not be available at the time of unwrapping.

Therefore, we refactor the unwrapping bits from the enhancer.

The code is generated with: [aas-core-codegen fe5e529]

[aas-core-codegen fe5e529]: https://github.com/aas-core-works/aas-core-codegen/commit/fe5e529