Fix error message in xmlization on XML declaration
When we encountered XML declarations, we threw exceptions with
uninformative messages. Namely, we expect the reader to be moved to the
content, but most users omitted to read that in the documentation.

We refine the error message for this particular situation, and hint at
`MoveToContent` method on the reader.

Fixes #24.

Generated with [aas-core-codegen 28fbd98].

[aas-core-codegen 28fbd98]: https://github.com/aas-core-works/aas-core-codegen/commit/28fbd98