Check explicitly for invalid model types
We explicitly check during the JSON de-serialization that model types
correspond to the expected model types. We need to be particularly
careful with concrete classes without descendants with a mandatory
``modelType``, as they do not necessarily require a model type for
de-serialization, but the specs mandate it for reasons of backward
compatibility.

The code corresponds to [aas-core-codegen b537610b], and the test data
corresponds to [aas-core3.0-testgen 9e523511c].

This is related to the issue [aas-core3.0-python #32], which discovered
the problematic in the first place.

[aas-core-codegen b537610b]: https://github.com/aas-core-works/aas-core-codegen/commit/b537610b
[aas-core3.0-testgen 9e523511c]: https://github.com/aas-core-works/aas-core3.0-testgen/commit/9e523511c
[aas-core3.0-python #32]: https://github.com/aas-core-works/aas-core3.0-python/issues/32